+++ Front/src/app/services/categorie.service.ts
@@ -4,6 +4,8 @@ import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 
 import { Categorie } from '../modeles/categorie';
+import { LoginService } from './login.service';
+import { Router } from '@angular/router';
 
 @Injectable({
   providedIn: 'root'
@@ -16,14 +18,24 @@ export class CategorieService {
   // La liste observable que l'on rend visible partout dans l'application
   availableCategorie$: BehaviorSubject<Categorie[]> = new BehaviorSubject(this.availableCategorie);
 
-  constructor(private httpClient: HttpClient) { }
+  constructor(private httpClient: HttpClient,
+              private loginService: LoginService,
+              private router: Router) { }
 
   getCategories(): Observable<Categorie[]> {
-    return this.httpClient.get<Categorie[]>('http://localhost:8080/categories/get/categories');
+    if (this.loginService.logged) {
+      return this.httpClient.get<Categorie[]>('http://localhost:8080/categories/get/categories');
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
   public getCategorie(): Observable<Categorie[]> {
-    return this.httpClient.get<Categorie[]>('http://localhost:8080/categories/get/categories');
+    if (this.loginService.logged) {
+      return this.httpClient.get<Categorie[]>('http://localhost:8080/categories/get/categories');
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
   public publishCategories() {
@@ -56,12 +68,16 @@ export class CategorieService {
    * @param newCategorie la nouvelle Categorie à créer
    */
   public createCategorie(newCategorie: Categorie) {
-    this.httpClient.post<Categorie>('http://localhost:8080/categories/create', newCategorie).subscribe(
-      createCategorie => {
-        this.availableCategorie.push(createCategorie);
-        this.availableCategorie$.next(this.availableCategorie);
-      }
-    );
+    if (this.loginService.logged) {
+      this.httpClient.post<Categorie>('http://localhost:8080/categories/create', newCategorie).subscribe(
+        createCategorie => {
+          this.availableCategorie.push(createCategorie);
+          this.availableCategorie$.next(this.availableCategorie);
+        }
+      );
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
   /**
@@ -69,12 +85,16 @@ export class CategorieService {
    * @param categorie la Categorie à mettre à jour
    */
   public updateCategorie(categorie: Categorie) {
-    this.httpClient.put<Categorie>(`http://localhost:8080/categories/update/${categorie.idCategorie}`, categorie).subscribe(
-      updateCategorie => {
-        this.availableCategorie.splice(this.availableCategorie.indexOf(categorie), 1, updateCategorie);
-        this.availableCategorie$.next(this.availableCategorie);
-      }
-    );
+    if (this.loginService.logged) {
+      this.httpClient.put<Categorie>(`http://localhost:8080/categories/update/${categorie.idCategorie}`, categorie).subscribe(
+        updateCategorie => {
+          this.availableCategorie.splice(this.availableCategorie.indexOf(categorie), 1, updateCategorie);
+          this.availableCategorie$.next(this.availableCategorie);
+        }
+      );
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
   /**
@@ -83,14 +103,18 @@ export class CategorieService {
    * @param idCategorie la Categorie à supprimer
    */
   supprimerCategorie(idCategorie: number): Categorie[] {
-    this.httpClient.delete('http://localhost:8080/categories/delete/' + idCategorie).subscribe(
-          () => { console.log('suppression categorie OK : ', idCategorie);
-                  this.availableCategorie = this.availableCategorie.filter( categorie => categorie.idCategorie !== idCategorie ).slice();
-                  this.availableCategorie$.next(this.availableCategorie);
-              },
-          (error) => console.log('suppression categorie pb : ', error)
-      );
-    return this.availableCategorie;
+    if (this.loginService.logged) {
+      this.httpClient.delete('http://localhost:8080/categories/delete/' + idCategorie).subscribe(
+            () => { console.log('suppression categorie OK : ', idCategorie);
+                    this.availableCategorie = this.availableCategorie.filter( categorie => categorie.idCategorie !== idCategorie ).slice();
+                    this.availableCategorie$.next(this.availableCategorie);
+                },
+            (error) => console.log('suppression categorie pb : ', error)
+        );
+      return this.availableCategorie;
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
 }
