+++ Front/src/app/services/disponibilite.service.ts
@@ -4,6 +4,8 @@ import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 
 import { Disponibilite } from '../modeles/disponibilite';
+import { LoginService } from './login.service';
+import { Router } from '@angular/router';
 
 @Injectable({
   providedIn: 'root'
@@ -16,10 +18,16 @@ export class DisponibiliteService {
   // La liste observable que l'on rend visible partout dans l'application
   availableDisponibilite$: BehaviorSubject<Disponibilite[]> = new BehaviorSubject(this.availableDisponibilite);
 
-  constructor(private httpClient: HttpClient) { }
+  constructor(private httpClient: HttpClient,
+              private loginService: LoginService,
+              private router: Router) { }
 
   public getDisponibilites(): Observable<Disponibilite[]> {
-    return this.httpClient.get<Disponibilite[]>('http://localhost:8080/disponibilite/get/disponibilites');
+    if (this.loginService.logged) {
+      return this.httpClient.get<Disponibilite[]>('http://localhost:8080/disponibilite/get/disponibilites');
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
   public publishDisponibilites() {
@@ -52,12 +60,16 @@ export class DisponibiliteService {
    * @param newDisponibilite la nouvelle disponibilite à créer
    */
   public createDisponibilite(newDisponibilite: Disponibilite) {
-    this.httpClient.post<Disponibilite>('http://localhost:8080/disponibilite/create', newDisponibilite).subscribe(
-      createDisponibilite => {
-        this.availableDisponibilite.push(createDisponibilite);
-        this.availableDisponibilite$.next(this.availableDisponibilite);
-      }
-    );
+    if (this.loginService.logged) {
+      this.httpClient.post<Disponibilite>('http://localhost:8080/disponibilite/create', newDisponibilite).subscribe(
+        createDisponibilite => {
+          this.availableDisponibilite.push(createDisponibilite);
+          this.availableDisponibilite$.next(this.availableDisponibilite);
+        }
+      );
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
 
@@ -66,12 +78,16 @@ export class DisponibiliteService {
    * @param disponibilite la disponibilite à mettre à jour
    */
   public updateDisponibilite(disponibilite: Disponibilite) {
-    this.httpClient.put<Disponibilite>(`http://localhost:8080/disponibilite/update/${disponibilite.idDisponibilite}`, disponibilite).subscribe(
-      updateDisponibilite => {
-        this.availableDisponibilite.splice(this.availableDisponibilite.indexOf(disponibilite), 1, updateDisponibilite);
-        this.availableDisponibilite$.next(this.availableDisponibilite);
-      }
-    );
+    if (this.loginService.logged) {
+      this.httpClient.put<Disponibilite>(`http://localhost:8080/disponibilite/update/${disponibilite.idDisponibilite}`, disponibilite).subscribe(
+        updateDisponibilite => {
+          this.availableDisponibilite.splice(this.availableDisponibilite.indexOf(disponibilite), 1, updateDisponibilite);
+          this.availableDisponibilite$.next(this.availableDisponibilite);
+        }
+      );
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 
   /**
@@ -80,13 +96,17 @@ export class DisponibiliteService {
    * @param idDisponibilite de la disponibilite à supprimer
    */
   supprimerDisponibilite(idDisponibilite: number): Disponibilite[] {
-    this.httpClient.delete('http://localhost:8080/disponibilite/delete/' + idDisponibilite).subscribe(
-          () => { console.log('suppression disponibilité OK : ', idDisponibilite);
-              },
-          (error) => console.log('suppression disponibilité pb : ', error)
-      );
-    this.availableDisponibilite = this.availableDisponibilite.filter( disponibilite => disponibilite.idDisponibilite !== idDisponibilite ).slice();
-    this.availableDisponibilite$.next(this.availableDisponibilite);
-    return this.availableDisponibilite;
+    if (this.loginService.logged) {
+      this.httpClient.delete('http://localhost:8080/disponibilite/delete/' + idDisponibilite).subscribe(
+            () => { console.log('suppression disponibilité OK : ', idDisponibilite);
+                },
+            (error) => console.log('suppression disponibilité pb : ', error)
+        );
+      this.availableDisponibilite = this.availableDisponibilite.filter( disponibilite => disponibilite.idDisponibilite !== idDisponibilite ).slice();
+      this.availableDisponibilite$.next(this.availableDisponibilite);
+      return this.availableDisponibilite;
+    } else {
+      this.router.navigate(['connexion']);
+    }
   }
 }
